Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325313 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 9f47aaa46710d632d472e89a419a3af42a65008b

+++ src/components/org/apache/jmeter/visualizers/Spline3.java
@@ -455,7 +455,7 @@ public class Spline3
         for (int i = 0; i < width; i++)
         {
             plot[i] =
-                (int) Math.round(((y[i] - min) * (height - 1)) / (max - min));
+                Math.round(((y[i] - min) * (height - 1)) / (max - min));
         }
         return plot;
     }
