+++ src/core/org/apache/jmeter/engine/StandardJMeterEngine.java
@@ -411,7 +411,7 @@ public class StandardJMeterEngine implements JMeterEngine, JMeterThreadMonitor,
                   notifier);
             threads[i].setThreadNum(i);
             threads[i].setInitialContext(JMeterContextService.getContext());
-            threads[i].setInitialDelay((int) (perThreadDelay * (float) i));
+            threads[i].setInitialDelay((int) (perThreadDelay * i));
             threads[i].setThreadName(groupName + " " + (groupCount) + "-" + (i + 1));
 
             scheduleThread(threads[i], group);
